requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -o requirements.txt requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.11
    # via discord-py
aiosignal==1.3.2
    # via aiohttp
altgraph==0.17.4
    # via pyinstaller
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via aiohttp
audioop-lts==0.2.1
    # via discord-py
discord-py==2.5.2
    # via -r requirements.in
frozenlist==1.6.2
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via yarl
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via
    #   pyinstaller
    #   pyinstaller-hooks-contrib
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.5
    # via -r requirements.in
pydantic-core==2.33.2
    # via pydantic
pyinstaller==6.14.0
    # via -r requirements.in
pyinstaller-hooks-contrib==2025.4
    # via pyinstaller
python-dotenv==1.1.0
    # via -r requirements.in
setuptools==80.9.0
    # via
    #   pyinstaller
    #   pyinstaller-hooks-contrib
tabulate==0.9.0
    # via -r requirements.in
typing-extensions==4.14.0
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
yarl==1.20.0
    # via aiohttp
